Extract remote keep-alive code into its own component
