--- src/Workspaces/Core/Portable/Rename/IRemoteRenameKeepAliveSession.cs
@@ -1,13 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System;
-
-namespace Microsoft.CodeAnalysis.Rename
-{
-    internal interface IRemoteRenameKeepAliveSession : IDisposable
-    {
-
-    }
-}
