+++ src/Workspaces/Remote/Core/ServiceDescriptors.cs
@@ -60,6 +60,7 @@ namespace Microsoft.CodeAnalysis.Remote
             (typeof(IRemoteSemanticClassificationService), null),
             (typeof(IRemoteDocumentHighlightsService), null),
             (typeof(IRemoteEncapsulateFieldService), typeof(IRemoteEncapsulateFieldService.ICallback)),
+            (typeof(IRemoteKeepAliveService), null),
             (typeof(IRemoteRenamerService), typeof(IRemoteRenamerService.ICallback)),
             (typeof(IRemoteConvertTupleToStructCodeRefactoringService), typeof(IRemoteConvertTupleToStructCodeRefactoringService.ICallback)),
             (typeof(IRemoteSymbolFinderService), typeof(IRemoteSymbolFinderService.ICallback)),
