+++ src/Workspaces/Remote/Core/xlf/RemoteWorkspacesResources.ja.xlf
@@ -82,6 +82,11 @@
         <target state="translated">継承の余白</target>
         <note />
       </trans-unit>
+      <trans-unit id="FeatureName_KeepAlive">
+        <source>Keep alive service</source>
+        <target state="new">Keep alive service</target>
+        <note />
+      </trans-unit>
       <trans-unit id="FeatureName_LegacySolutionEventsAggregation">
         <source>Solution Events</source>
         <target state="translated">ソリューション イベント</target>
