+++ src/Workspaces/Remote/Core/xlf/RemoteWorkspacesResources.ko.xlf
@@ -82,6 +82,11 @@
         <target state="translated">상속 여백</target>
         <note />
       </trans-unit>
+      <trans-unit id="FeatureName_KeepAlive">
+        <source>Keep alive service</source>
+        <target state="new">Keep alive service</target>
+        <note />
+      </trans-unit>
       <trans-unit id="FeatureName_LegacySolutionEventsAggregation">
         <source>Solution Events</source>
         <target state="translated">솔루션 이벤트</target>
